Add docker-compose file (docker-compose-all.yml) which, in addition, starts a container for the front-end.
- docker-compose-all.yml is like docker-compose.yml except that it
addtionally sets up a container for the backend. It compiles the
front-end project and starts a dev server, reachable at localhost:8100.
*It was necessary to change the frontend/angular.json such that it
listens on 0.0.0.0, this means that running the webserver outside of
the container will serve for other hosts!*
- for both docker-compose files the backend-server now waits for the
database to fully boot before doing anything. This prevents connection
errors if the database needs a little longer.